refactor: split GGUF quantization into hardware backends and schemes
