+++ python/sglang/srt/hardware_backend/npu/quantization/gguf_kernels.py
@@ -1,0 +1,318 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Optional
+
+import gguf
+import torch
+from gguf import GGMLQuantizationType as WeightType
+from gguf import dequantize as gguf_dequantize
+
+if TYPE_CHECKING:
+    from sglang.srt.layers.moe import MoeRunnerConfig
+    from sglang.srt.layers.moe.token_dispatcher import StandardDispatchOutput
+    from sglang.srt.layers.quantization.base_config import QuantizationConfig
+
+UNQUANTIZED_TYPES = {WeightType.F32, WeightType.F16, WeightType.BF16}
+STANDARD_QUANT_TYPES = {
+    WeightType.Q4_0,
+    WeightType.Q4_1,
+    WeightType.Q5_0,
+    WeightType.Q5_1,
+    WeightType.Q8_0,
+    WeightType.Q8_1,
+}
+KQUANT_TYPES = {
+    WeightType.Q2_K,
+    WeightType.Q3_K,
+    WeightType.Q4_K,
+    WeightType.Q5_K,
+    WeightType.Q6_K,
+}
+IMATRIX_QUANT_TYPES = {
+    WeightType.IQ1_M,
+    WeightType.IQ1_S,
+    WeightType.IQ2_XXS,
+    WeightType.IQ2_XS,
+    WeightType.IQ2_S,
+    WeightType.IQ3_XXS,
+    WeightType.IQ3_S,
+    WeightType.IQ4_XS,
+    WeightType.IQ4_NL,
+}
+DEQUANT_TYPES = STANDARD_QUANT_TYPES | KQUANT_TYPES | IMATRIX_QUANT_TYPES
+
+
+def ggml_dequantize_ascend(
+    qweight: torch.Tensor,
+    qweight_type: int,
+    rows: int,
+    cols: int,
+    dtype: torch.dtype,
+) -> torch.Tensor:
+    """Dequantize GGML quantized weights for NPU.
+
+    Uses gguf library's reference implementation which supports all GGML formats
+    and is guaranteed to be correct. The dequantization runs on CPU during model
+    loading, then the dequantized weights are transferred to NPU for inference.
+    """
+    qweight_cpu = qweight.cpu().numpy()
+    dequant_np = gguf_dequantize(qweight_cpu, qweight_type)
+    result = torch.from_numpy(dequant_np).to(dtype=dtype, device=qweight.device)
+    result = result.reshape(rows, cols)
+    return result
+
+
+class GGUFAscendLinearKernel:
+    def __init__(self, quant_config: Optional["QuantizationConfig"] = None):
+        self.quant_config = quant_config
+        self.params_dtype: torch.dtype | None = None
+
+    def process_weights_after_loading(self, layer: torch.nn.Module) -> None:
+        qweight_type = layer.qweight_type.weight_type
+        if not (qweight_type in UNQUANTIZED_TYPES or qweight_type in DEQUANT_TYPES):
+            raise ValueError(
+                f"Unsupported GGUF quantization type {WeightType(qweight_type)} in layer."
+            )
+        self._pre_dequantize_weights(layer)
+
+    def _pre_dequantize_weights(self, layer: torch.nn.Module) -> None:
+        qweight = layer.qweight
+        qweight_type = layer.qweight_type.weight_type
+
+        if qweight_type in UNQUANTIZED_TYPES and qweight.dtype in (
+            torch.float16,
+            torch.bfloat16,
+            torch.float32,
+        ):
+            layer.dequantized_weight = qweight
+            return
+
+        shard_id = getattr(qweight, "shard_id", None)
+        has_shard_offset = hasattr(qweight, "shard_offset_map")
+
+        if shard_id and has_shard_offset:
+            shard_id = ["q", "k", "v"] if "q" in shard_id else shard_id
+            dequant_shards = []
+            for idx in shard_id:
+                start, end, offset = qweight.shard_offset_map[idx]
+                shard_qtype = layer.qweight_type.shard_weight_type[idx]
+                shard_data = qweight[start:end, :offset].contiguous()
+
+                block_size, type_size = gguf.GGML_QUANT_SIZES[shard_qtype]
+                shape = (
+                    shard_data.shape[0],
+                    shard_data.shape[1] // type_size * block_size,
+                )
+                dequant = ggml_dequantize_ascend(
+                    shard_data, shard_qtype, *shape, self.params_dtype
+                )
+                dequant_shards.append(dequant)
+
+            dequant_weight = torch.cat(dequant_shards, dim=0)
+        else:
+            block_size, type_size = gguf.GGML_QUANT_SIZES[qweight_type]
+            shape = (qweight.shape[0], qweight.shape[1] // type_size * block_size)
+            dequant_weight = ggml_dequantize_ascend(
+                qweight, qweight_type, *shape, self.params_dtype
+            )
+
+        layer.dequantized_weight = dequant_weight
+
+        if hasattr(layer, "qweight"):
+            del layer.qweight
+        if hasattr(layer, "qweight_type"):
+            del layer.qweight_type
+
+    def apply(
+        self,
+        layer: torch.nn.Module,
+        x: torch.Tensor,
+        bias: Optional[torch.Tensor] = None,
+    ) -> torch.Tensor:
+        weight = layer.dequantized_weight
+        out = x @ weight.T
+        if bias is not None:
+            out.add_(bias)
+        return out
+
+
+class GGUFAscendEmbeddingKernel(GGUFAscendLinearKernel):
+    def embedding(self, layer: torch.nn.Module, x: torch.Tensor) -> torch.Tensor:
+        return torch.embedding(layer.dequantized_weight, x)
+
+
+class GGUFAscendMoEKernel:
+    def __init__(self, quant_config: Optional["QuantizationConfig"] = None):
+        self.quant_config = quant_config
+        self.moe_runner_config: Optional["MoeRunnerConfig"] = None
+        self.params_dtype: torch.dtype | None = None
+
+    def create_moe_runner(
+        self, layer: torch.nn.Module, moe_runner_config: "MoeRunnerConfig"
+    ) -> None:
+        self.moe_runner_config = moe_runner_config
+
+    def process_weights_after_loading(self, layer: torch.nn.Module) -> None:
+        if hasattr(layer, "materialize_gguf_weights"):
+            layer.materialize_gguf_weights()
+
+        w13_qweight = layer.w13_qweight
+        w13_qtype = layer.w13_qweight_type.weight_type
+
+        if w13_qtype not in UNQUANTIZED_TYPES:
+            num_experts = w13_qweight.shape[0]
+            w13_dequant_list = []
+
+            block_size, type_size = gguf.GGML_QUANT_SIZES[w13_qtype]
+
+            for e in range(num_experts):
+                qweight_cpu = w13_qweight[e].cpu().numpy()
+                rows = w13_qweight[e].shape[0]
+                cols = w13_qweight[e].shape[1] // type_size * block_size
+
+                dequant_np = gguf_dequantize(qweight_cpu.flatten(), w13_qtype)
+                dequant = (
+                    torch.from_numpy(dequant_np)
+                    .to(dtype=self.params_dtype, device=w13_qweight.device)
+                    .reshape(rows, cols)
+                    .transpose(-1, -2)
+                    .contiguous()
+                )
+                w13_dequant_list.append(dequant)
+
+            w13_full = torch.stack(w13_dequant_list, dim=0)
+            layer.register_buffer("w13_dequant", w13_full, persistent=False)
+        else:
+            layer.register_buffer("w13_dequant", w13_qweight.data, persistent=False)
+
+        w2_qweight = layer.w2_qweight
+        w2_qtype = layer.w2_qweight_type.weight_type
+
+        if w2_qtype not in UNQUANTIZED_TYPES:
+            num_experts = w2_qweight.shape[0]
+            w2_dequant_list = []
+
+            block_size, type_size = gguf.GGML_QUANT_SIZES[w2_qtype]
+
+            for e in range(num_experts):
+                qweight_cpu = w2_qweight[e].cpu().numpy()
+                rows = w2_qweight[e].shape[0]
+                cols = w2_qweight[e].shape[1] // type_size * block_size
+
+                dequant_np = gguf_dequantize(qweight_cpu.flatten(), w2_qtype)
+                dequant = (
+                    torch.from_numpy(dequant_np)
+                    .to(dtype=self.params_dtype, device=w2_qweight.device)
+                    .reshape(rows, cols)
+                    .transpose(-1, -2)
+                    .contiguous()
+                )
+                w2_dequant_list.append(dequant)
+
+            w2_full = torch.stack(w2_dequant_list, dim=0)
+            layer.register_buffer("w2_dequant", w2_full, persistent=False)
+        else:
+            layer.register_buffer("w2_dequant", w2_qweight.data, persistent=False)
+
+        if hasattr(layer, "w2_qweight"):
+            del layer.w2_qweight
+        if hasattr(layer, "w13_qweight"):
+            del layer.w13_qweight
+
+    def apply(
+        self,
+        layer: torch.nn.Module,
+        dispatch_output: "StandardDispatchOutput",
+    ):
+        from sglang.srt.distributed.communication_op import (
+            tensor_model_parallel_all_gather,
+        )
+        from sglang.srt.layers.moe.token_dispatcher import StandardCombineInput
+
+        x = dispatch_output.hidden_states
+        topk_output = dispatch_output.topk_output
+        topk_weights, topk_ids, _ = topk_output
+
+        if not (hasattr(layer, "w13_dequant") and hasattr(layer, "w2_dequant")):
+            raise RuntimeError(
+                "GGUF MoE on NPU requires pre-dequantization (FusedMoE fix). "
+                "Please report if this occurs."
+            )
+
+        w13 = layer.w13_dequant
+        w2 = layer.w2_dequant
+
+        num_experts = w13.shape[0]
+        tp_size = getattr(layer, "moe_tp_size", 1)
+
+        original_dtype = x.dtype
+        num_tokens = x.shape[0]
+        top_k = topk_ids.shape[1]
+
+        topk_ids = topk_ids.to(torch.int32)
+        topk_weights = topk_weights.to(x.dtype)
+
+        row_idx_len = num_tokens * top_k
+        row_idx = (
+            torch.arange(0, row_idx_len, dtype=torch.int32, device=x.device)
+            .view(top_k, -1)
+            .permute(1, 0)
+            .contiguous()
+        )
+
+        sorted_hidden_states, expanded_row_idx, expanded_expert_idx = (
+            torch.ops.npu.npu_moe_init_routing(
+                x, row_idx=row_idx, expert_idx=topk_ids, active_num=num_tokens
+            )
+        )
+
+        expert_tokens = torch.ops.npu.npu_moe_compute_expert_tokens(
+            expanded_expert_idx, num_experts
+        )
+        expert_tokens = expert_tokens.to(torch.int64)
+
+        hidden_states = torch.ops.npu.npu_grouped_matmul(
+            x=[sorted_hidden_states],
+            weight=[w13],
+            split_item=2,
+            group_list_type=0,
+            group_type=0,
+            group_list=expert_tokens,
+            output_dtype=original_dtype,
+        )[0]
+
+        hidden_states = torch.ops.npu.npu_swiglu(hidden_states)
+
+        if tp_size > 1:
+            hidden_states = tensor_model_parallel_all_gather(hidden_states, dim=-1)
+
+        hidden_states = torch.ops.npu.npu_grouped_matmul(
+            x=[hidden_states],
+            weight=[w2],
+            split_item=2,
+            group_list_type=0,
+            group_type=0,
+            group_list=expert_tokens,
+            output_dtype=original_dtype,
+        )[0]
+
+        final_hidden_states = torch.ops.npu.npu_moe_finalize_routing(
+            hidden_states,
+            skip1=None,
+            skip2=None,
+            bias=None,
+            scales=topk_weights,
+            expanded_src_to_dst_row=expanded_row_idx,
+            export_for_source_row=topk_ids,
+        )
+
+        if tp_size > 1:
+            final_hidden_states = tensor_model_parallel_all_gather(
+                final_hidden_states, dim=-1
+            )
+
+        final_hidden_states = final_hidden_states.to(dtype=original_dtype)
+
+        return StandardCombineInput(hidden_states=final_hidden_states)
